Created TransactionModal component
