+++ savewise-client/src/App.css
@@ -39,3 +39,21 @@
 .regularColor {
     color: #1c1c1c;
 }
+
+.modalBg {
+    border-radius: 5px;
+    padding-left: 10px;
+    min-height: 200px;
+    width: 400px;
+    background-color:   rgba(219, 219, 219, 1);
+    margin: 20px auto 0 auto;
+    padding: 0 15px 20px 15px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.buttonGroupModal {
+    margin-top: 20px;
+}
