+++ savewise-client/src/components/Dashboard/Dashboard.module.scss
@@ -11,19 +11,6 @@
     box-shadow:  16px 16px 32px #cecece,
              -16px -16px 32px #e8e8e8;
 }
-.modalBg {
-    border-radius: 5px;
-    padding-left: 10px;
-    min-height: 400px;
-    width: 400px;
-    background-color:   rgba(219, 219, 219, 1);
-    margin: 20px auto 0 auto;
-    padding: 0 15px 20px 15px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-}
 .itemsContainer {
     width: 90%;
     height: 100%;
