+++ savewise-client/src/components/TransactionAdmin/TransactionAdmin.module.scss
@@ -1,19 +1,3 @@
-.modalBg {
-    border-radius: 5px;
-    padding-left: 10px;
-    min-height: 400px;
-    width: 400px;
-    background-color:   rgba(219, 219, 219, 1);
-    margin: 20px auto 0 auto;
-    padding: 0 15px 20px 15px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-between;
-}
-.buttonGroup {
-    margin-top: 20px;
-}
 .transactionItem {
     width: 98%;
     padding: 3px 6px;
