+++ savewise-client/src/components/TransactionAdmin/TransactionAdmin.tsx
@@ -1,26 +1,18 @@
-import React, { ChangeEvent, useContext, useEffect, useState } from "react";
-import { Category } from "../../common/objects/categories";
+import React, { useContext, useEffect, useState } from "react";
 import { Transaction, TransactionForm } from "../../common/objects/transactions";
 import { LoginContext } from "../../common/context/LoginContext";
-import { GetCategoriesInput } from "../../services/category-service";
 import {
-    CategoryService,
     TransactionService,
     UtilService,
 } from "../../services";
-import InputLabel from "@material-ui/core/InputLabel";
-import MenuItem from "@material-ui/core/MenuItem";
-import FormControl from "@material-ui/core/FormControl";
 import IconButton from "@material-ui/core/IconButton";
 import EditIcon from "@material-ui/icons/Edit";
 import DeleteIcon from "@material-ui/icons/Delete";
 import { Link } from "react-router-dom";
 import ArrowBackIosIcon from "@material-ui/icons/ArrowBackIos";
 import Modal from "@material-ui/core/Modal";
-import TextField from "@material-ui/core/TextField";
 import Button from "@material-ui/core/Button";
 import styles from "./TransactionAdmin.module.scss";
-import Select from "@material-ui/core/Select";
 import { KeyboardDatePicker } from "@material-ui/pickers";
 import { Status } from "../../common/objects/response";
 import SearchIcon from '@material-ui/icons/Search';
@@ -28,6 +20,7 @@ import CancelIcon from '@material-ui/icons/Cancel';
 import { SnackbarContext } from "../../common/context/SnackbarContext";
 import { constants } from "../../common/objects/constants";
 import { CategoryTypesId } from "../../common/objects/CategoryTypesId";
+import TransactionModal from "../TransactionModal/TransactionModal";
 
 interface SearchByDates {
     fromDate: string,
@@ -48,27 +41,11 @@ export default function TransactionAdmin() {
         description: "",
     });
     const [loading, setLoading] = useState(true);
-    const [userCategories, setUserCategories] = useState<Category[]>([]);
     const [showSearch, setShowSearch] = useState<boolean>(false);
     const [searchForm, setSearchForm] = useState<SearchByDates>({ fromDate: UtilService.today(), toDate: UtilService.tomorrow() });
 
     useEffect(() => {
         const id: number = login.login?.id as number;
-        const categoriesOptions: GetCategoriesInput = {
-            includeAmounts: false,
-            startDate: "",
-            endDate: "",
-        };
-        CategoryService.getCategories(id, categoriesOptions)
-            .then((rsp) => {
-                if (rsp.status.success) {
-                    setUserCategories(rsp.categories);
-                } else {
-                    console.log(`Error: ${rsp.status.errorMessage}`);
-                }
-            })
-            .catch((e) => console.log(`ERROR: ${e}`));
-
         TransactionService.GetTransactions(id, undefined, undefined, 10)
             .then((rsp) => {
                 if (rsp.status.success) {
@@ -81,7 +58,7 @@ export default function TransactionAdmin() {
             .catch((e) => console.log(`ERROR: ${e}`));
 
         return () => {
-            setUserCategories([]);
+
         };
     }, [login]);
 
@@ -213,75 +190,15 @@ export default function TransactionAdmin() {
         setOpenConfirmDelete(!openConfirmDelete);
     };
 
-    const handleInputChange = (event: ChangeEvent<HTMLInputElement>) => {
-        if (event.target.name === "amount") {
-            setTransactionForm({
-                ...transactionForm,
-                [event.target.name]: Number(event.target.value),
-            });
+    const onSave = (savedTransaction: Transaction) => {
+        let updatedTransactions: Transaction[];
+        if (transactions.find((t) => t.id === savedTransaction.id)) {
+            updatedTransactions = transactions.map((t) => t.id === savedTransaction.id ? savedTransaction : t);
         } else {
-            setTransactionForm({
-                ...transactionForm,
-                [event.target.name]: event.target.value,
-            });
+            updatedTransactions = [...transactions, savedTransaction];
         }
-    };
-
-    const handleSelectChange = (event: React.ChangeEvent<{ value: unknown }>) => {
-        setTransactionForm({
-            ...transactionForm,
-            categoryId: event.target.value as number,
-        });
-    };
-
-    const handleDateChange = (date: Date | null) => {
-        setTransactionForm({
-            ...transactionForm,
-            date: date ? UtilService.formatDate(date) : "",
-        });
-    };
-
-    const renderUserCategories = (userCategories: Category[]) => {
-        if (userCategories.length) {
-            return userCategories.map((category) => {
-                return (
-                    <MenuItem key={category.id} value={category.id}>
-                        {category.name}
-                    </MenuItem>
-                );
-            });
-        }
-    };
-
-    const onSave = () => {
-        const transaction: Transaction = {
-            userId: login.login?.id as number,
-            category: userCategories.find(
-                (c) => c.id === transactionForm.categoryId
-            ) as Category,
-            amount: transactionForm.amount as number,
-            date: transactionForm.date,
-            description: transactionForm.description,
-            id: transactionForm.id,
-        };
-        TransactionService.SaveTransaction(transaction)
-            .then((rsp) => {
-                if (rsp.status.success) {
-                    handleToggleModal();
-                    let updatedTransactions: Transaction[];
-                    if (transactions.find((t) => t.id === rsp.transaction.id)) {
-                        updatedTransactions = UtilService.sortTransactionByDate(transactions.map((t) =>t.id === rsp.transaction.id ? rsp.transaction : t))
-                    } else {
-                        updatedTransactions = [...transactions, rsp.transaction];
-                    }
-                    const sortedTransactions = UtilService.sortTransactionByDate(updatedTransactions);
-                    setTransactions(sortedTransactions);
-                    setSnackbarInfo({ severity: "success", message: "Transaction saved successfully!" })
-                } else {
-                    setSnackbarInfo({ severity: "error", message: rsp.status.errorMessage });
-                }
-            })
-            .catch((e) => setSnackbarInfo({ severity: "error", message: e }));
+        const sortedTransactions = UtilService.sortTransactionByDate(updatedTransactions);
+        setTransactions(sortedTransactions);
     };
 
     const onConfirmDelete = () => {
@@ -302,77 +219,8 @@ export default function TransactionAdmin() {
         }
     };
 
-    const modalBody = (
-        <div className={styles.modalBg}>
-            {transactionForm.id ? 
-            (<h1 className="titleColor">Edit transaction</h1>) 
-            : 
-            (<h1 className="titleColor">Create transaction</h1>)}
-            <TextField
-                id="outlined-basic"
-                onChange={handleInputChange}
-                label="Transaction Name"
-                variant="outlined"
-                name="description"
-                value={transactionForm.description}
-                fullWidth
-            />
-            <TextField
-                id="outlined-basic"
-                onChange={handleInputChange}
-                label="Amount (€)"
-                variant="outlined"
-                name="amount"
-                type="number"
-                value={transactionForm.amount ? transactionForm.amount : ""}
-                style={{ marginTop: "10px" }}
-                placeholder="€"
-                fullWidth
-            />
-            <FormControl
-                style={{ marginTop: "10px" }}
-                variant="outlined"
-                fullWidth
-            >
-                <InputLabel id="category">Category</InputLabel>
-                <Select
-                    labelId="category"
-                    id="categorySelect"
-                    value={transactionForm.categoryId ? transactionForm.categoryId : ""}
-                    onChange={handleSelectChange}
-                >
-                    {renderUserCategories(userCategories)}
-                </Select>
-            </FormControl>
-            <KeyboardDatePicker
-                inputVariant="outlined"
-                margin="normal"
-                id="date-picker-dialog"
-                label="Date"
-                format="dd/MM/yyyy"
-                value={UtilService.formatStringDatePicker(transactionForm.date)}
-                onChange={handleDateChange}
-                KeyboardButtonProps={{
-                    "aria-label": "change date",
-                }}
-                style={{ marginTop: "10px" }}
-                fullWidth
-            />
-            <div className={styles.buttonGroup}>
-                <Button
-                    variant="contained"
-                    color="primary"
-                    type="submit"
-                    onClick={onSave}
-                >
-                    Save
-                </Button>
-            </div>
-        </div>
-    );
-
     const confirmDeleteBody = (
-        <div className={styles.modalBg}>
+        <div className="modalBg">
             <h1>Delete transaction {transactionForm.description}</h1>
             <p>This is irreversible. Are you sure?</p>
             <Button
@@ -418,13 +266,7 @@ export default function TransactionAdmin() {
                         {renderTransactions(transactions)}
                 </div>)
             }
-            <Modal
-                open={openModal}
-                onClose={() => handleToggleModal()}
-                aria-labelledby="add-transaction"
-            >
-                {modalBody}
-            </Modal>
+            <TransactionModal conditionToShow={openModal} handleVisibility={() => setOpenModal(!openModal)} handleSave={(t: Transaction) => onSave(t)} />
             <Modal
                 open={openConfirmDelete}
                 onClose={() => handleToggleDeleteModal()}
